+++ activestorage/test/test_helper.rb
@@ -188,6 +188,8 @@ class User < ActiveRecord::Base
   accepts_nested_attributes_for :highlights_attachments, allow_destroy: true
 end
 
+class SpecialUser < User; end
+
 class Group < ActiveRecord::Base
   has_one_attached :avatar
   has_many :users, autosave: true